+++ src/lib/action.ts
@@ -27,7 +27,7 @@ const isOpenProps = (props: AnyObject): props is OpenProps => {
  * Used to orchestrate mounting of Clerk components in a host Svelte application.
  * Components are rendered into a specific DOM node using mount/unmount methods provided by the Clerk class.
  */
-export const clerkUI: Action<HTMLDivElement, MountProps | OpenProps> = (node, props) => {
+export const clerkHostRenderer: Action<HTMLDivElement, MountProps | OpenProps> = (node, props) => {
 	if (isMountProps(props)) {
 		props.mount(node, props.props);
 	} else if (isOpenProps(props)) {
@@ -49,5 +49,3 @@ export const clerkUI: Action<HTMLDivElement, MountProps | OpenProps> = (node, pr
 		}
 	};
 };
-
-export default clerkUI;