+++ src/lib/client/ClerkProvider.svelte
@@ -11,7 +11,6 @@
 		type LoadClerkJsScriptOptions
 	} from '@clerk/shared/loadClerkJsScript';
 	import { goto } from '$app/navigation';
-	import { removeNetlifyCacheBustParam } from '$lib/utils/netlifyCacheBust';
 
 	const {
 		children,
@@ -63,7 +62,6 @@
 
 	$effect(() => {
 		untrack(() => {
-			removeNetlifyCacheBustParam();
 			loadClerk();
 		});
 	});