+++ src/lib/client/interactive/GoogleOneTap.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import { clerkUI } from '$lib/action.js';
+	import { clerkHostRenderer } from '$lib/action.js';
 	import type { GoogleOneTapProps } from '@clerk/types';
 	import ClerkLoaded from '$lib/client/control/ClerkLoaded.svelte';
 
@@ -9,7 +9,7 @@
 <ClerkLoaded>
 	{#snippet children(clerk)}
 		<div
-			use:clerkUI={{
+			use:clerkHostRenderer={{
 				open: clerk.openGoogleOneTap,
 				close: clerk.closeGoogleOneTap,
 				props: $state.snapshot(props)