+++ src/lib/server/currentUser.ts
@@ -1,8 +1,11 @@
-import type { AuthObject, User } from '@clerk/backend';
+import type { User } from '@clerk/backend';
+import type { SignedInAuthObject, SignedOutAuthObject } from '@clerk/backend/internal';
 import { clerkClient } from './clerkClient.js';
+import { deprecated } from '@clerk/shared/deprecated';
 
-export function createCurrentUser(auth: AuthObject) {
+export function createCurrentUser(auth: SignedInAuthObject | SignedOutAuthObject) {
 	return async (): Promise<User | null> => {
+	  deprecated('currentUser()', 'Use `clerkClient.users.getUser(id)` instead.')
 		if (!auth.userId) {
 			return null;
 		}